requirements.txt
pandas
openpyxl
mysql-connector
matplotlib
selenium
prettytable
